Update onyx.c
onyx is going to be a different kind of database language. haven't quite gotten to the groundbreaking, earthshattering innovational aspects as yet however, because I have only barely just started in on the core functionality (which as you can see, looks quite alot more suited to evolving into some kindof an obscure line-parsing library. in any event, and whatever the case, do not steal my name! OnyxBase is MINE. :) My.